Site Setup: Add basic "Webpage" framework
Initially we setup and verified that we could talk to the db and
output some PHP logic.

Now we need to create acutal "web pages" which requires a bit of
structure. There is a lot of meta information that needs to be
handled in a real site.

This adds all the basic pieces of the puzzle:
 * Site Runner - Allows different "pages" to be run
 * HTML Pages - Associates .phtml templates to a "Page Type"
 * CSS - Allows things to look pretty
 * JS - Allows things to move pretty

In the end, there's a single "Dev Page" that gets displayed.

It should have a header & an article phtml template, css, js
and should dump the db data to verify that's still working.

Note: This framework is hacked off of my website. I know there
are better libraries, this is just a sandbox I built & like to
play in.