Site Setup: Abstract out DB Logic
We were connecting to the DB and grabbing the right data but it
was a major hack job that could only do that.

We're probably going to want to do several DB queries and we'd
like to have access to the data in a nice iterable format. This
creates a "DBUtils" library to do that basic setup stuff.

Also, now that the data is coming in correctly we can update the
html to display our DB data correctly instead of just dumping the
values on the backend.

Finally this does some basic css stuff to setup the page better.