Site Setup: Accepted POSTed comment form
Now that we can see the data in a nice table we want to be able to
add to it.

This adds a basic form to the page which takes an name and a comment
and it adds it to the comments table. The page reloads after so the
user can see their new comment.

Because queries are getting more complex, db tables now have their
own files to handle their logic. DBUtils is now just a general way
of managing the db.

We're accepting user data so all the values are parameterized, which
should help protect against SQL injection.

There's a small hoop we had to jump through in looking up the userid
off of the "name" but that was intended. Now we have parameterized
query and insert logic.

Some cleanup was also handled here. Mainly changing the db table
name from `user` to `users` and removing a bunch of invalid includes.